move AgentEntry to entry_agent.go
